温度显示bug

+++ app/src/main/java/com/moko/beaconxpro/utils/BeaconXParser.java
@@ -2,7 +2,6 @@ package com.moko.beaconxpro.utils;
 
 import android.text.TextUtils;
 
-
 import com.moko.beaconxpro.entity.BeaconXAxis;
 import com.moko.beaconxpro.entity.BeaconXDevice;
 import com.moko.beaconxpro.entity.BeaconXTH;
@@ -102,7 +101,7 @@ public class BeaconXParser {
         BeaconXTH beaconXTH = new BeaconXTH();
         int rssi_0m = Integer.parseInt(data.substring(2, 4), 16);
         beaconXTH.rangingData = (byte) rssi_0m + "";
-        beaconXTH.temperature = MokoUtils.getDecimalFormat("0.0").format(Integer.parseInt(data.substring(6, 10), 16) * 0.1f);
+        beaconXTH.temperature = MokoUtils.getDecimalFormat("0.0").format(((short) Integer.parseInt(data.substring(6, 10), 16)) * 0.1f);
         beaconXTH.humidity = MokoUtils.getDecimalFormat("0.0").format(Integer.parseInt(data.substring(10, 14), 16) * 0.1f);
         return beaconXTH;
     }
